+++ R/dimensio-package.R
@@ -2,10 +2,15 @@
 #' \tabular{ll}{
 #'  **Package:** \tab dimensio \cr
 #'  **Type:** \tab Package \cr
-#'  **Version:** \tab 0.10.1 \cr
+#'  **Version:** \tab 0.11.0 \cr
 #'  **License:** \tab GPL-3 \cr
 #'  **Zenodo:** \tab \doi{10.5281/zenodo.4478530} \cr
 #' }
+#'
+#' @section Package options:
+#'  \pkg{dimensio} uses the following [options()] to configure behavior:
+#'  * `dimensio.verbose`: a [`logical`] scalar. Should \R report extra
+#'    information on progress? Defaults to [interactive()].
 #'
 #' @author
 #' **Full list of authors and contributors** (alphabetic order)