+++ R/summary.R
@@ -5,16 +5,16 @@ NULL
 # CA ===========================================================================
 #' @export
 #' @method summary CA
-summary.CA <- function(object, ..., margin = 1, active = TRUE, sup = TRUE,
-                       rank = 3) {
+summary.CA <- function(object, ..., axes = c(1, 2), margin = 1,
+                       active = TRUE, sup = TRUE, rank = NULL) {
   ## Get data
-  values <- build_summary(object, margin = margin, rank = rank,
+  values <- build_summary(object, axes = axes, margin = margin, rank = rank,
                           active = active, sup = sup)
 
   .SummaryCA(
     data = object@data,
-    eigenvalues = as.matrix(values$eigenvalues),
-    results = as.matrix(values$results),
+    eigenvalues = values$eigenvalues,
+    results = values$results,
     supplement = values$supplement,
     margin = as.integer(margin)
   )
@@ -25,19 +25,43 @@ summary.CA <- function(object, ..., margin = 1, active = TRUE, sup = TRUE,
 #' @aliases summary,CA-method
 setMethod("summary", c(object = "CA"), summary.CA)
 
+#' @export
+#' @rdname describe
+#' @aliases describe,CA-method
+setMethod(
+  f = "describe",
+  signature = signature(x = "CA"),
+  definition = function(x, ...) {
+    row_sup <- x@rows@supplement
+    col_sup <- x@columns@supplement
+
+    sup_txt <- tr_(" (+ %d supplementary)")
+    row_txt <- if (any(row_sup)) sprintf(sup_txt, sum(row_sup)) else ""
+    col_txt <- if (any(col_sup)) sprintf(sup_txt, sum(col_sup)) else ""
+
+    cat(
+      sprintf(tr_("* Row variable: %d categories%s."), sum(!row_sup), row_txt),
+      sprintf(tr_("* Column variable: %d categories%s."), sum(!col_sup), col_txt),
+      ...,
+      sep = "\n"
+    )
+    invisible(x)
+  }
+)
+
 # PCA ==========================================================================
 #' @export
 #' @method summary PCA
-summary.PCA <- function(object, ..., margin = 1, active = TRUE, sup = TRUE,
-                        rank = 3) {
+summary.PCA <- function(object, ..., axes = c(1, 2), margin = 1,
+                        active = TRUE, sup = TRUE, rank = NULL) {
   ## Get data
-  values <- build_summary(object, margin = margin, rank = rank,
+  values <- build_summary(object, axes = axes, margin = margin, rank = rank,
                           active = active, sup = sup)
 
   .SummaryPCA(
     data = object@data,
-    eigenvalues = as.matrix(values$eigenvalues),
-    results = as.matrix(values$results),
+    eigenvalues = values$eigenvalues,
+    results = values$results,
     supplement = values$supplement,
     margin = as.integer(margin)
   )
@@ -48,9 +72,58 @@ summary.PCA <- function(object, ..., margin = 1, active = TRUE, sup = TRUE,
 #' @aliases summary,PCA-method
 setMethod("summary", c(object = "PCA"), summary.PCA)
 
-build_summary <- function(object, margin, rank = 3,
+#' @export
+#' @rdname describe
+#' @aliases describe,PCA-method
+setMethod(
+  f = "describe",
+  signature = signature(x = "PCA"),
+  definition = function(x, ...) {
+    row_sup <- x@rows@supplement
+    col_sup <- x@columns@supplement
+
+    sup_txt <- tr_(" (+ %d supplementary)")
+    row_txt <- if (any(row_sup)) sprintf(sup_txt, sum(row_sup)) else ""
+    col_txt <- if (any(col_sup)) sprintf(sup_txt, sum(col_sup)) else ""
+
+    if (is_centered(x)) {
+      var_center <- tr_("* Variables were shifted to be zero centered.")
+    } else {
+      var_center <- tr_("* Variables were NOT shifted to be zero centered.")
+    }
+    if (is_scaled(x)) {
+      var_scale <- tr_("* Variables were scaled to unit variance.")
+    } else {
+      var_scale <- tr_("* Variables were NOT scaled to unit variance.")
+    }
+
+    cat(
+      sprintf(tr_("* %d individuals%s."), sum(!row_sup), row_txt),
+      sprintf(tr_("* %d variables%s."), sum(!col_sup), col_txt),
+      var_center,
+      var_scale,
+      ...,
+      sep = "\n"
+    )
+    invisible(x)
+  }
+)
+
+# Helpers ======================================================================
+build_summary <- function(object, axes, margin, rank = NULL,
                           active = TRUE, sup = TRUE,
                           prefix = "F") {
+  ## Validation
+  arkhe::assert_filled(axes)
+  arkhe::assert_type(axes, "numeric")
+
+  ## /!\ Backward compatibility /!\
+  if (!is.null(rank)) {
+    axes <- seq_len(rank)
+    msg <- "'rank' argument is deprecated, use 'axes' instead."
+    warning(msg, call. = FALSE)
+  }
+
   ## Get data
   eig <- get_eigenvalues(object)
   inertia <- get_distances(object, margin = margin)
@@ -69,15 +142,14 @@ build_summary <- function(object, margin, rank = 3,
   }
 
   ## Bind columns
-  rank <- min(rank, dim(object))
-  dim_keep <- seq_len(rank)
-  values <- vector(mode = "list", length = rank)
-  for (j in dim_keep) {
-    v <- data.frame(coord[[j]], contrib[[j]], cos2[[j]])
-    names(v) <- paste0(prefix, j, c("_coord", "_contrib", "_cos2"))
+  values <- vector(mode = "list", length = length(axes))
+  for (j in axes) {
+    v <- cbind(coord[[j]], contrib[[j]], cos2[[j]])
+    colnames(v) <- paste0(prefix, j, c("_coord", "_contrib", "_cos2"))
     values[[j]] <- v
   }
-  values <- data.frame(inertia = inertia, values)
+  values <- do.call(cbind, values)
+  values <- cbind(inertia = inertia, values)
   if (inherits(object, "PCA")) colnames(values)[1] <- "dist"
   rownames(values) <- rownames(coord)
 
@@ -85,13 +157,17 @@ build_summary <- function(object, margin, rank = 3,
   is_sup <- coord$.sup
   if (!active && !sup) active <- TRUE
   if (!active) {
-    values <- values[is_sup, ]
+    values <- values[is_sup, , drop = FALSE]
     is_sup <- is_sup[is_sup]
   }
   if (!sup) {
-    values <- values[!is_sup, ]
+    values <- values[!is_sup, , drop = FALSE]
     is_sup <- is_sup[!is_sup]
   }
 
-  list(eigenvalues = eig, results = values, supplement = is_sup)
+  list(
+    eigenvalues = as.matrix(eig),
+    results = as.matrix(values),
+    supplement = is_sup
+  )
 }